Extract filtering functionals
