+++ torchaudio/functional/__init__.py
@@ -1,40 +1,42 @@
 from .functional import (
-    allpass_biquad,
     amplitude_to_DB,
     angle,
-    band_biquad,
-    bandpass_biquad,
-    bandreject_biquad,
-    bass_biquad,
-    biquad,
-    DB_to_amplitude,
     complex_norm,
     compute_deltas,
-    contrast,
     create_dct,
     create_fb_matrix,
-    dither,
-    dcshift,
-    deemph_biquad,
+    DB_to_amplitude,
     detect_pitch_frequency,
-    equalizer_biquad,
-    flanger,
-    gain,
     griffinlim,
-    highpass_biquad,
-    lfilter,
-    lowpass_biquad,
     magphase,
     mask_along_axis,
     mask_along_axis_iid,
     mu_law_encoding,
     mu_law_decoding,
-    overdrive,
     phase_vocoder,
-    phaser,
-    riaa_biquad,
     sliding_window_cmn,
     spectrogram,
+)
+from .filtering import (
+    allpass_biquad,
+    band_biquad,
+    bandpass_biquad,
+    bandreject_biquad,
+    bass_biquad,
+    biquad,
+    contrast,
+    dither,
+    dcshift,
+    deemph_biquad,
+    equalizer_biquad,
+    flanger,
+    gain,
+    highpass_biquad,
+    lfilter,
+    lowpass_biquad,
+    overdrive,
+    phaser,
+    riaa_biquad,
     treble_biquad,
     vad,
 )
